backend/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-dotenv
